Notebook preparation for PANGEOS UQ summer school
The idea here is to make a python package with all the functionality
that was stored before in some random python files. This makes
installing and importing it easier.

I have also extended some of the examples to deal with uncertainty
caused by atmospheric correction. This still needs some work
and updating of a WiP notebook.

Most of the inversions have been done using MCMC because I find it
quite educational and straightforward to implement.